feat(VCST-5126): refactor UCP handoff cart restoration logic to shared module
