+++ client-app/router/routes/checkout.ts
@@ -1,6 +1,5 @@
-import { USER_ID_LOCAL_STORAGE } from "@/core/constants";
-import { setGlobals } from "@/core/globals";
 import { ROUTES } from "./constants";
+import { applyUcpHandoffBuyer, restoreUcpHandoffCart } from "./ucp-handoff";
 import type { NavigationGuardNext, RouteLocationNormalized, RouteRecordName, RouteRecordRaw } from "vue-router";
 
 const Checkout = () => import("@/pages/checkout/index.vue");
@@ -10,13 +9,6 @@ const PaymentResult = () => import("@/pages/checkout/payment-result.vue");
 const Payment = () => import("@/pages/checkout/payment.vue");
 const Review = () => import("@/pages/checkout/review.vue");
 const Shipping = () => import("@/pages/checkout/shipping.vue");
-
-type UcpHandoffRestoreType = {
-  cartId: string;
-  buyerId?: string;
-};
-
-const UCP_HANDOFF_STORAGE_PREFIX = "ucp-handoff:";
 
 function handleBeforeEnter(
   from: RouteLocationNormalized,
@@ -29,84 +21,6 @@ function handleBeforeEnter(
   } else {
     next();
   }
-}
-
-function getUcpHandoffStorageKey(ucpSession: string): string {
-  return `${UCP_HANDOFF_STORAGE_PREFIX}${ucpSession}`;
-}
-
-function applyUcpHandoffBuyer(buyerId?: string): void {
-  if (!buyerId) {
-    return;
-  }
-
-  localStorage.setItem(USER_ID_LOCAL_STORAGE, buyerId);
-  setGlobals({ userId: buyerId });
-}
-
-// eslint-disable-next-line sonarjs/function-return-type
-function getCachedUcpHandoffRestore(ucpSession: string): UcpHandoffRestoreType | undefined {
-  const value = sessionStorage.getItem(getUcpHandoffStorageKey(ucpSession));
-  let result: UcpHandoffRestoreType | undefined;
-
-  if (value) {
-    try {
-      result = JSON.parse(value) as UcpHandoffRestoreType;
-    } catch {
-      sessionStorage.removeItem(getUcpHandoffStorageKey(ucpSession));
-    }
-  }
-
-  return result;
-}
-
-function cacheUcpHandoffRestore(ucpSession: string, value: UcpHandoffRestoreType): void {
-  sessionStorage.setItem(getUcpHandoffStorageKey(ucpSession), JSON.stringify(value));
-}
-
-async function restoreUcpHandoffCart(ucpSession: string): Promise<UcpHandoffRestoreType> {
-  const headers = new Headers({
-    "content-type": "application/json",
-    accept: "application/json",
-  });
-
-  const response = await fetch("/ucp/v1/internal/handoff/restore", {
-    method: "POST",
-    credentials: "omit",
-    cache: "no-store",
-    headers,
-    body: JSON.stringify({ ucp_session: ucpSession }),
-  });
-
-  if (!response.ok) {
-    const cached = getCachedUcpHandoffRestore(ucpSession);
-
-    if (cached) {
-      return cached;
-    }
-
-    const details = await response.text().catch(() => "");
-    throw new Error(`Unable to restore UCP handoff session. Status: ${response.status}. ${details}`);
-  }
-
-  const payload = (await response.json()) as {
-    checkout?: {
-      cart_id?: string;
-      buyer?: { id?: string };
-      cart?: { id?: string; buyer_id?: string };
-    };
-  };
-  const cartId = payload.checkout?.cart_id ?? payload.checkout?.cart?.id;
-  const buyerId = payload.checkout?.buyer?.id ?? payload.checkout?.cart?.buyer_id;
-
-  if (!cartId) {
-    throw new Error("UCP handoff session does not include cart id.");
-  }
-
-  const restore = { cartId, buyerId };
-  cacheUcpHandoffRestore(ucpSession, restore);
-
-  return restore;
 }
 
 export const checkoutRoutes: RouteRecordRaw[] = [
